+++ tests/integration/drift/compile_only/clickhouse/02_total_sales_by_country.sql
@@ -1,5 +1,5 @@
 SELECT "Countries"."countryname" AS "Sales Country Name", CAST(round(SUM("Sales"."salesamount"), 2) AS Nullable(Decimal(18, 2))) AS "Total Sales"
 FROM "orionbelt_1"."sales" AS "Sales"
-LEFT JOIN "orionbelt_1"."clients" AS "Clients" ON ("Sales"."salesclient" = "Clients"."clientid")
-LEFT JOIN "orionbelt_1"."countries" AS "Countries" ON ("Clients"."clientcountryid" = "Countries"."countryid")
+LEFT JOIN "orionbelt_1"."clients" AS "Clients" ON "Sales"."salesclient" = "Clients"."clientid"
+LEFT JOIN "orionbelt_1"."countries" AS "Countries" ON "Clients"."clientcountryid" = "Countries"."countryid"
 GROUP BY "Countries"."countryname"