+++ tests/integration/drift/compile_only/databricks/12_total_sales_country_in_filter.sql
@@ -1,5 +1,5 @@
 SELECT CAST(SUM(`Sales`.`salesamount`) AS DECIMAL(18, 2)) AS `Total Sales`
 FROM ``.`orionbelt_1`.`sales` AS `Sales`
-LEFT JOIN ``.`orionbelt_1`.`clients` AS `Clients` ON (`Sales`.`salesclient` = `Clients`.`clientid`)
-LEFT JOIN ``.`orionbelt_1`.`countries` AS `Countries` ON (`Clients`.`clientcountryid` = `Countries`.`countryid`)
-WHERE (`Countries`.`countryname` IN ('Croatia', 'Austria'))
+LEFT JOIN ``.`orionbelt_1`.`clients` AS `Clients` ON `Sales`.`salesclient` = `Clients`.`clientid`
+LEFT JOIN ``.`orionbelt_1`.`countries` AS `Countries` ON `Clients`.`clientcountryid` = `Countries`.`countryid`
+WHERE `Countries`.`countryname` IN ('Croatia', 'Austria')